Moved corAppealTasks into redux action
